pyproject.toml
[project]
name = "ig-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "cenzwong", email = "cenzth@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]

[build-system]
requires = ["uv_build>=0.9.28,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mkdocs>=1.6.1",
    "mkdocs-jupyter>=0.25.1",
    "mkdocs-material>=9.7.1",
    "mkdocstrings[python]>=1.0.2",
    "pytest>=8.0.0",
    "ipykernel>=6.29.0",
]
